get rid of residual deprecation warnings in DataFormats/SiStripCommon
