--- DataFormats/SiStripCommon/test/plugins/modules.cc
@@ -1,28 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "DataFormats/SiStripCommon/test/plugins/test_SiStripFecKey.h"
-DEFINE_FWK_MODULE(testSiStripFecKey);
-
-#include "DataFormats/SiStripCommon/test/plugins/examples_SiStripFecKey.h"
-DEFINE_FWK_MODULE(examplesSiStripFecKey);
-
-#include "DataFormats/SiStripCommon/test/plugins/perf_SiStripFecKey.h"
-DEFINE_FWK_MODULE(perfSiStripFecKey);
-
-#include "DataFormats/SiStripCommon/test/plugins/test_SiStripFedKey.h"
-DEFINE_FWK_MODULE(testSiStripFedKey);
-
-#include "DataFormats/SiStripCommon/test/plugins/test_SiStripNullKey.h"
-DEFINE_FWK_MODULE(testSiStripNullKey);
-
-#include "DataFormats/SiStripCommon/test/plugins/test_SiStripKey.h"
-DEFINE_FWK_MODULE(testSiStripKey);
-
-#include "DataFormats/SiStripCommon/test/plugins/test_SiStripEnumsAndStrings.h"
-DEFINE_FWK_MODULE(testSiStripEnumsAndStrings);
-
-#include "DataFormats/SiStripCommon/test/plugins/test_SiStripHistoTitle.h"
-DEFINE_FWK_MODULE(testSiStripHistoTitle);
-
-#include "DataFormats/SiStripCommon/test/plugins/test_Template.h"
-DEFINE_FWK_MODULE(test_Template);
